requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
distlib==0.3.6
filelock==3.12.0
idna==3.4
pipenv==2023.4.29
platformdirs==3.5.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
urllib3==2.0.4
virtualenv==20.23.0
virtualenv-clone==0.5.7
